PressureSensorCode
This should be close to final for the race code. Further testing with water needs to be done before it is complete. The function "pressure_function()" outputs the velocity and depth of the sub. If there is something I'm missing for calling the function for the race code, let me know, I am not proficient at Arduino yet.